Agrego ArticuloRepository

+++ PRACTICAS/Factura/Facturas/Facturas/Dominio/Articulo.cs
@@ -10,7 +10,7 @@ namespace Facturas.Dominio
     {
         public int IdArticulo { get; set; }
         public string Nombre { get; set; }
-        public int PrecioUnitario { get; set; }
+        public double PrecioUnitario { get; set; }
         public override string? ToString()
         {
             return Nombre + "[" + IdArticulo + "]";
